Move argument counts to int32_t.

+++ libndtypes/equal.c
@@ -54,9 +54,9 @@ ndt_common_equal(const ndt_t *t, const ndt_t *u)
 }
 
 static int
-function_types_equal(const ndt_t *t, const ndt_t *u, int64_t nargs)
+function_types_equal(const ndt_t *t, const ndt_t *u, int32_t nargs)
 {
-    int64_t i;
+    int32_t i;
 
     assert(t->tag == Function && u->tag == Function);
